+++ server/src/enum/entity-partition.enum.ts
@@ -1,0 +1,9 @@
+export enum EntityPartition {
+  FAMILIAR = '1 << 0',
+  BULLET = '1 << 1',
+  TEAR = '1 << 2',
+  ENEMY = '1 << 3',
+  PICKUP = '1 << 4',
+  PLAYER = '1 << 5',
+  EFFECT = '1 << 6'
+}