+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 Completion for EntityFlag Enum.
 Completion for EntityPartition Enum.
+Completion for GameStateFlag Enum.
 ## 0.0.14
 ### Added
 Completion for EntityCollisionClass Enum.