+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 Completion for EntityFlag Enum.
 Completion for EntityPartition Enum.
 Completion for GameStateFlag Enum.
+Completion for GridCollisionClass Enum.
 ## 0.0.14
 ### Added
 Completion for EntityCollisionClass Enum.