+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Completion for EntityFlag Enum.
 Completion for EntityPartition Enum.
 Completion for GameStateFlag Enum.
 Completion for GridCollisionClass Enum.
+Completion for GridEntityType Enum.
 ## 0.0.14
 ### Added
 Completion for EntityCollisionClass Enum.